fix: Fixes #127 (#448)
This change hardcodes the path to `bash` when it is used within Docker.
Ideally, we should be able to use `env` to locate `bash`, it seems
that some container execution platforms do not support it.

Invoking `bash` from the development, CI, or Ubuntu full installs will
continue to use the `env` pattern.

Signed-off-by: Elijah Zupancic <e.zupancic@f5.com>
Co-authored-by: ajimenez <ajimenez@lendingpoint.com>